requirements.txt
sqlalchemy==0.9.4
pytest==2.5.2,
pytest-cov==1.6
colorama==0.3.1
termcolor==1.1.0
logbook==0.7.0
pyyaml==3.11